requirements.txt
gitpython
hvac
ansible
apscheduler
JPype1
six
dataset
